+++ code_pylint.py
@@ -1,6 +1,8 @@
 import os
 import random
 import sys
+from datetime import date
+import math
 
 from pylint import __version__
 from pylint.lint import Run
@@ -9,37 +11,36 @@ MIN_NOTE = 8.20
 
 UNWATCHED_ERRORS = ['fixme', 'trailing-whitespace', 'import-error', 'missing-final-newline']
 
+EFFECTIVE_DATE = date(2023, 1, 31)
+WEEKLY_DECREASE = 0.03
 
 MAX_ERROR_BY_TYPE = {
-                     "wrong-spelling-in-comment": 400,
-                     "wrong-spelling-in-docstring": 600,
-                     'invalid-name': 780,
-                     'no-else-return': 49,
-                     'consider-using-f-string': 77,
+                     "wrong-spelling-in-comment": 335,
+                     "wrong-spelling-in-docstring": 352,
+                     'invalid-name': 729,
+                     'no-else-return': 28,
+                     'consider-using-f-string': 69,
                      'no-member': 9,
-                     'inconsistent-return-statements': 6,
-                     'unused-variable': 42,
-                     'arguments-differ': 12,
-                     'too-many-locals': 73,
-                     'line-too-long': 23,
-                     'unused-argument': 43,
-                     'too-many-arguments': 63,
-                     'line-too-long errors': 32,
+                     'inconsistent-return-statements': 4,
+                     'unused-variable': 32,
+                     'arguments-differ': 58,
+                     'too-many-locals': 74,
+                     'unused-argument': 41,
+                     'too-many-arguments': 62,
+                     'line-too-long': 19,
                      'consider-using-enumerate': 22,
                      'too-many-branches': 27,
                      'too-many-statements': 18,
-                     'super-init-not-called': 12,
+                     'super-init-not-called': 13,
                      'no-name-in-module': 5,
-                     'abstract-method': 35,
-                     'empty-docstring': 15,
-                     'duplicate-code': 7,
-                     'no-self-use': 16,
-                     'arguments-renamed': 3,
-                     'too-many-ancestors': 8,
-                     'expression-not-assigned': 1,
-                     'non-parent-init-called': 6,
+                     'abstract-method': 36,
+                     'duplicate-code': 9,
+                     'arguments-renamed': 8,
+                     'too-many-ancestors': 29,
                      'too-few-public-methods': 2,
-                     'too-many-public-methods': 10,
+                     # 'expression-not-assigned': 1,
+                     'non-parent-init-called': 6,
+                     'too-many-public-methods': 11,
                      'use-implicit-booleaness-not-comparison': 8,
                      'too-many-instance-attributes': 10,
                      'protected-access': 4,
@@ -47,89 +48,54 @@ MAX_ERROR_BY_TYPE = {
                      'unspecified-encoding': 1,
                      'too-many-function-args': 4,
                      'too-many-nested-blocks': 7,
-                     'attribute-defined-outside-init': 5,
+                     'attribute-defined-outside-init': 6,
                      'too-many-return-statements': 1,
-                     'consider-merging-isinstance': 0,
                      'cyclic-import': 4,
-                     'consider-iterating-dictionary': 0,
                      'raise-missing-from': 2,
                      'no-else-raise': 3,
-                     'no-else-continue': 4,
+                     'no-else-continue': 3,
                      'undefined-variable': 6,  # 2 when gmsh is fixed
                      'no-else-break': 4,
-                     'unnecessary-list-index-lookup': 4,
-                     'simplifiable-if-expression': 3,
-                     'redefined-builtin': 3,
                      'broad-except': 1,
                      "broad-exception-caught": 1,
                      'too-many-boolean-expressions': 3,
                      'too-many-lines': 3,
                      'redundant-keyword-arg': 3,
-                     'no-value-for-parameter': 1,
                      'c-extension-no-member': 2,
-                     'access-member-before-definition': 1,
+                     # 'access-member-before-definition': 1,
                      'modified-iterating-list': 2,
                      'consider-using-with': 1,
-                     'consider-using-get': 2,
                      'unnecessary-dunder-call': 2,
                      'unnecessary-lambda': 2,
-                     'eval-used': 2,
                      'chained-comparison': 2,
                      'missing-module-docstring': 2,
-                     'unbalanced-tuple-unpacking': 2,
-                     'bad-staticmethod-argument': 1,
                      'consider-using-generator': 1,
-                     'use-maxsplit-arg': 1,
-                     'wildcard-import': 1,
                      'cell-var-from-loop': 1,
                      'import-outside-toplevel': 1,
                      'unsubscriptable-object': 1,
-                     # No tolerance errors
-                     'unidiomatic-typecheck': 0,
-                     'unexpected-special-method-signature': 0,
-                     'bare-except': 0,
-                     'function-redefined': 0,
-                     'superfluous-parens': 0,
-                     'unnecessary-comprehension': 0,
-                     'unused-wildcard-import': 0,
-                     'wrong-import-order': 0,
-                     'ungrouped-imports': 0,
-                     'assignment-from-none': 0,
-                     'non-iterator-returned': 0,
-                     'consider-using-max-builtin': 0,
-                     'consider-using-min-builtin': 0,
-                     'format-string-without-interpolation': 0,
-                     'bad-indentation': 0,
-                     'implicit-str-concat': 0,
-                     'return-in-init': 0,
-                     'redefined-outer-name': 0,
-                     'no-self-argument': 0,
-                     'consider-using-from-import': 0,
-                     'duplicate-string-formatting-argument': 0,
-                     'wrong-import-position': 0,
-                     'singleton-comparison': 0,
-                     'unreachable': 0,
-                     'consider-using-in': 0,
-                     'unused-import': 0
                      }
 
-print('pylint version: ', __version__)
+ERRORS_WITHOUT_TIME_DECREASE = []
 
-f = open(os.devnull, 'w')
+print("pylint version: ", __version__)
+
+time_decrease_coeff = 1 - (date.today() - EFFECTIVE_DATE).days / 7.0 * WEEKLY_DECREASE
+
+f = open(os.devnull, "w")
 
 old_stdout = sys.stdout
 sys.stdout = f
 
-results = Run(['volmdlr', '--output-format=json', '--reports=no'], do_exit=False)
+results = Run(["volmdlr", "--output-format=json", "--reports=no"], do_exit=False)
 # `exit` is deprecated, use `do_exit` instead
 sys.stdout = old_stdout
 
 PYLINT_OBJECTS = True
-if hasattr(results.linter.stats, 'global_note'):
+if hasattr(results.linter.stats, "global_note"):
     pylint_note = results.linter.stats.global_note
     PYLINT_OBJECT_STATS = True
 else:
-    pylint_note = results.linter.stats['global_note']
+    pylint_note = results.linter.stats["global_note"]
     PYLINT_OBJECT_STATS = False
 
 
@@ -143,36 +109,46 @@ error_over_ratchet_limit = False
 if PYLINT_OBJECT_STATS:
     stats_by_msg = results.linter.stats.by_msg
 else:
-    stats_by_msg = results.linter.stats['by_msg']
+    stats_by_msg = results.linter.stats["by_msg"]
 
 for error_type, number_errors in stats_by_msg.items():
     if error_type not in UNWATCHED_ERRORS:
-        max_errors = MAX_ERROR_BY_TYPE.get(error_type, 0)
+        base_errors = MAX_ERROR_BY_TYPE.get(error_type, 0)
+
+        if error_type in ERRORS_WITHOUT_TIME_DECREASE:
+            max_errors = base_errors
+        else:
+            max_errors = math.ceil(base_errors * time_decrease_coeff)
+
+        time_decrease_effect = base_errors - max_errors
+        # print('time_decrease_effect', time_decrease_effect)
 
         if number_errors > max_errors:
             error_detected = True
-            print(f'\nFix some {error_type} errors: {number_errors}/{max_errors}')
+            print(
+                f"\nFix some {error_type} errors: {number_errors}/{max_errors} "
+                f"(time effect: {time_decrease_effect} errors)")
 
             messages = extract_messages_by_type(error_type)
-            messages_to_show = sorted(random.sample(messages, min(40, len(messages))),
-                                      key=lambda m: (m.path, m.line))
+            messages_to_show = sorted(random.sample(messages, min(30, len(messages))), key=lambda m: (m.path, m.line))
             for message in messages_to_show:
-                print(f'{message.path} line {message.line}: {message.msg}')
+                print(f"{message.path} line {message.line}: {message.msg}")
         elif number_errors < max_errors:
-            print(f'\nYou can lower number of {error_type} to {number_errors} (actual {max_errors})')
+            print(f"\nYou can lower number of {error_type} to {number_errors+time_decrease_effect}"
+                  f" (actual {base_errors})")
 
+for error_type in MAX_ERROR_BY_TYPE:
+    if error_type not in stats_by_msg:
+        print(f"You can delete {error_type} entry from MAX_ERROR_BY_TYPE dict")
 
 if error_detected:
-    raise RuntimeError('Too many errors\nRun pylint volmdlr to get the errors')
+    raise RuntimeError("Too many errors\nRun pylint volmdlr to get the errors")
 
 if error_over_ratchet_limit:
-    raise RuntimeError('Please lower the error limits in code_pylint.py MAX_ERROR_BY_TYPE according to warnings above')
+    raise RuntimeError("Please lower the error limits in code_pylint.py MAX_ERROR_BY_TYPE according to warnings above")
 
-print('Pylint note: ', pylint_note)
-# if pylint_note > MIN_NOTE + RATCHET_NOTE:
-#     raise ValueError(f'MIN_NOTE in code_pylint.py is too low, increase to at least {MIN_NOTE + RATCHET_NOTE}, max {pylint_note}')
+print("Pylint note: ", pylint_note)
 if pylint_note < MIN_NOTE:
-    raise ValueError(f'Pylint not is too low: {pylint_note}, expected {MIN_NOTE}')
+    raise ValueError(f"Pylint not is too low: {pylint_note}, expected {MIN_NOTE}")
 
-print('You can increase MIN_NOTE in pylint to {} (actual: {})'.format(pylint_note,
-                                                                      MIN_NOTE))
+print("You can increase MIN_NOTE in pylint to {} (actual: {})".format(pylint_note, MIN_NOTE))