+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+* edges: remove attributes points from lines & linesegments for performance purpose
 
 ### Performance improvements
 * wires.py's 2D objects: chache bounding_rectangle results
@@ -129,7 +130,7 @@ abscissa(), point_angle_with_major_dir(), area(), rotation(), tranlation(), fram
 * cache variable self._polygon_point_belongs_100, to avoid recalculating each
 time we have to verify if a point is inside
 * Improvements in BSplineSurface3D.point3d_to_2d performance
-
+* Triangle3D serialization speed-up
 
 ### Refactorings
 * Basis2D, Basis3D, Frame2D, Frame3D: old_coordinates and new_coordinates method are now deprecated.